MicMods

Documentation for MicMods.

MicMods.LabeledParSetter

MicMods.LabeledParSetter

MicMods.ParSetter

MicMods.ParSetter

MicMods.ParSetter

MicMods.estimate_u0_from_cumresp

MicMods.extendrange

MicMods.find_inflection

MicMods.getpopt

MicMods.getpopt

MicMods.kinresp_exp

MicMods.kinresp_mic

MicMods.label_parsys

MicMods.label_parsys

MicMods.label_popt

MicMods.label_popt

MicMods.label_statesys

MicMods.label_statesys

MicMods.micfromcoef

MicMods.parindex

MicMods.replacetby0

MicMods.setpu

MicMods.setpu

MicMods.stateindex

MicMods.LabeledParSetter — Type

LabeledParSetter

Translates between a vector of optimization parameters and the two vectors of a system: its parameters and initial states.

MicMods.LabeledParSetter — Method

LabeledParSetter(
    systemt, 
    paropt = keys(systemt.searchranges_p), 
    stateopt = keys(systemt.searchranges_u0))

Construct a Parameter-Setter based on the subset of parameters and initial states.

MicMods.ParSetter — Type

ParSetter

Translates between a vector of optimization parameters and the two vectors of a system: its parameters and initial states.

MicMods.ParSetter — Method

ParSetter(parsys,statesys,paropt,stateopt)

Construct a Parameter-Setter based on the subset of parameters and initial states.

values

All arguments are of type AbstractVector{Num}

parsys,statesys: parameters and states of the system

paropt,stateopt: parameters and states in optimization vector

MicMods.ParSetter — Method

ParSetter(
    systemt, 
    paropt = keys(systemt.searchranges_p), 
    stateopt = keys(systemt.searchranges_u0))

Construct a Parameter-Setter based on the subset of parameters and initial states.

MicMods.estimate_u0_from_cumresp — Method

estimate_u0_from_cumresp(tinfl, popt0, systemt, chak21syn; ps = LabelledParSetter(systemt))

Obtain initial estimate of substrate s and biomass u by computing uptake from microbial parameters and respiration.

Arguments

tinfl: (numeric) start-time, after amendment

u0: labelled array of initial state with entries s and b

b0: microbial biomass at amendmend time 0. The default is obtained from chak21syn problem

t, r_tot: respiration rate measured at time t

r0, Y: representative phyisological state and yield for the period 0..tinfl.

Value

Tuple (u0, cr_tot_gr) of parameters and cumulated respiration.

MicMods.extendrange — Function

extendrange(lower,upper, frac = 0.1)

Modify lower and upper to so that the included range is extended by frac.

This is useful for settting plotting limits based on quantiles.

MicMods.find_inflection — Method

find_inflection(y,y)

Find the inflection point in the first increase where the slope is maximum.

Arguments

y,y: numeric vectors to fit relationship

Examples

true

MicMods.getpopt — Method

getpopt(ps::LabeledParSetter, prob)

Extract the parameter vector from prob.p and prob.u0.

Value: Numeric array.

MicMods.getpopt — Method

getpopt(ps::ParSetter, prob)

Extract the parameter vector from prob.p and prob.u0.

Value: Numeric array.

MicMods.kinresp_exp — Method

Fit exponential equation with lognormal priors on coefficients.

MicMods.kinresp_mic — Method

Fit the kinetic respiraiton model using microbial parameters

MicMods.label_parsys — Method

label_parsys(ps::ParSetter, popt)

Create an LVector from parameter vector

MicMods.label_parsys — Method

label_parsys(ps::LabeledParSetter, popt)

Create an LVector from parameter vector

MicMods.label_popt — Method

label_popt(ps::ParSetter, popt)

Create an LVector from parameter vector

MicMods.label_popt — Method

label_popt(ps::LabeledParSetter, popt)

Create an LVector from parameter vector

MicMods.label_statesys — Method

label_parsys(ps::ParSetter, popt)

Create an LVector from parameter vector

MicMods.label_statesys — Method

label_parsys(ps::LabeledParSetter, popt)

Create an LVector from parameter vector

MicMods.micfromcoef — Method

micfromcoef(params; λ = 0.9, YCO2 = 1.5)

Compute microbial parameters from coefficient regressions β0,β1,β2. See Wutzler11 eq. (2,3,4)

Value: Numeric array: [μmax, r0, x0, λ, YCO2]

MicMods.parindex — Method

parindex(ps::ParSetter, sym::Symbol)

Get the index corresponding to symbol in vector of system parameters.

Value: Integer or nothing

MicMods.replacetby0 — Method

replace "(t)" in sting vector by "_0"

To create parameter names, that MCMCChains.get can handle. get(chn, Symbol("b(t)") is not extracted.

MicMods.setpu — Method

setpu(ps::LabeledParSetter, popt, prob)
setpu(ps::LabeledParSetter, popt, p, u0)

Set values in tuple (prob.p, prob.u0) to popt.

Eltype of new p and u0 should correspond to type of popt, while vector type should correspond to original p and u0. If p is of type SVector{3,Float64} and eltype(popt) == Dual128, then new p is of type SVector{3,Dual128}.

Value: Tuple (p, u0)

MicMods.setpu — Method

setpu(ps::ParSetter, popt, prob)
setpu(ps::ParSetter, popt, p, u0)

Set values in tuple (prob.p, prob.u0) to popt.

Eltype of new p and u0 should correspond to type of popt, while vector type should correspond to original p and u0. If p is of type SVector{3,Float64} and eltype(popt) == Dual128, then new p is of type SVector{3,Dual128}.

Value: Tuple (p, u0)

MicMods.stateindex — Method

stateindex(ps::ParSetter, sym::Symbol)

Get the index corresponding to symbol in vector of system states.

Value: Integer or nothing